fix minor compil warning (signed/unsigned compar)

+++ eeschema/dialogs/dialog_fields_editor_global.cpp
@@ -456,7 +456,7 @@ public:
 
     void CollapseAll()
     {
-        for( int i = 0; i < m_rows.size(); ++i )
+        for( size_t i = 0; i < m_rows.size(); ++i )
         {
             if( m_rows[ i ].m_Flag == GROUP_EXPANDED )
                 CollapseRow( i );
